feat: migrate esi requests to openAPI

+++ memberaudit/managers/character_sections_2.py
@@ -5,11 +5,11 @@
 from __future__ import annotations
 
 import ast
+from http import HTTPStatus
 from typing import TYPE_CHECKING, Any, Dict, List, Set
 
-from bravado.exception import HTTPNotFound
-
 from django.db import models, transaction
+from esi.exceptions import HTTPClientError
 from esi.models import Token
 from eveuniverse.models import (
     EveAncestry,
@@ -63,11 +63,10 @@ class CharacterCorporationHistoryManager(GenericUpdateComplexObjMixin, models.Ma
 
     def _fetch_data_from_esi(self, character: Character) -> List[dict]:
         logger.info("%s: Fetching corporation history from ESI", character)
-        history = esi.client.Character.get_characters_character_id_corporationhistory(
+        history = esi.client.Character.GetCharactersCharacterIdCorporationhistory(
             character_id=character.eve_character.character_id,
-        ).results()
-
-        return history
+        ).result(use_etag=False)
+        return [obj.model_dump() for obj in history]
 
     def _update_or_create_objs(
         self, character: Character, esi_data: List[dict]
@@ -106,13 +105,12 @@ class CharacterDetailsManager(models.Manager):
             force_update=force_update,
         )
 
-    def _fetch_data_from_esi(self, character: Character):
+    def _fetch_data_from_esi(self, character: Character) -> Dict[str, Any]:
         logger.info("%s: Fetching character details from ESI", character)
-        details = esi.client.Character.get_characters_character_id(
+        details = esi.client.Character.GetCharactersCharacterId(
             character_id=character.eve_character.character_id,
-        ).results()
-
-        return details
+        ).result(use_etag=False)
+        return details.model_dump()
 
     def _update_or_create_objs(self, character: Character, details) -> Set[int]:
         description = details.get("description") or ""
@@ -182,11 +180,11 @@ class CharacterFwStatsManager(models.Manager):
     @fetch_token_for_character("esi-characters.read_fw_stats.v1")
     def _fetch_data_from_esi(self, character: Character, token: Token):
         logger.info("%s: Fetching FW stats from ESI", character)
-        stats = esi.client.Faction_Warfare.get_characters_character_id_fw_stats(
+        stats = esi.client.Faction_Warfare.GetCharactersCharacterIdFwStats(
             character_id=character.eve_character.character_id,
-            token=token.valid_access_token(),
-        ).results()
-        return stats
+            token=token,
+        ).result(use_etag=False)
+        return stats.model_dump()
 
     def _update_or_create_objs(self, character: Character, stats: dict):
         if faction_id := stats.get("faction_id"):
@@ -225,10 +223,10 @@ class CharacterImplantManager(models.Manager):
     @fetch_token_for_character("esi-clones.read_implants.v1")
     def _fetch_data_from_esi(self, character: Character, token: Token) -> List[int]:
         logger.info("%s: Fetching implants from ESI", character)
-        implant_type_ids = esi.client.Clones.get_characters_character_id_implants(
+        implant_type_ids = esi.client.Clones.GetCharactersCharacterIdImplants(
             character_id=character.eve_character.character_id,
-            token=token.valid_access_token(),
-        ).results()
+            token=token,
+        ).result(use_etag=False)
         return implant_type_ids
 
     def _update_or_create_objs(self, character: Character, implant_type_ids: List[int]):
@@ -268,11 +266,11 @@ class CharacterJumpCloneManager(models.Manager):
     @fetch_token_for_character("esi-clones.read_clones.v1")
     def _fetch_data_from_esi(self, character: Character, token: Token):
         logger.info("%s: Fetching jump clones from ESI", character)
-        jump_clones_info = esi.client.Clones.get_characters_character_id_clones(
+        jump_clones_info = esi.client.Clones.GetCharactersCharacterIdClones(
             character_id=character.eve_character.character_id,
-            token=token.valid_access_token(),
-        ).results()
-        return jump_clones_info
+            token=token,
+        ).result(use_etag=False)
+        return jump_clones_info.model_dump()
 
     # TODO: Replace delete & create with update
     @fetch_token_for_character("esi-universe.read_structures.v1")
@@ -387,11 +385,11 @@ class CharacterLocationManager(models.Manager):
     )
     def _fetch_data_from_esi(self, character: Character, token):
         logger.info("%s: Fetching location from ESI", character)
-        location_info = esi.client.Location.get_characters_character_id_location(
+        location_info = esi.client.Location.GetCharactersCharacterIdLocation(
             character_id=character.eve_character.character_id,
-            token=token.valid_access_token(),
-        ).results()
-        return location_info
+            token=token,
+        ).result(use_etag=False)
+        return location_info.model_dump()
 
     @fetch_token_for_character(
         ["esi-location.read_location.v1", "esi-universe.read_structures.v1"]
@@ -438,12 +436,11 @@ class CharacterLoyaltyEntryManager(GenericUpdateSimpleObjMixin, models.Manager):
     @fetch_token_for_character("esi-characters.read_loyalty.v1")
     def _fetch_data_from_esi(self, character: Character, token: Token) -> List[dict]:
         logger.info("%s: Fetching loyalty entries from ESI", character)
-        loyalty_entries = esi.client.Loyalty.get_characters_character_id_loyalty_points(
+        loyalty_entries = esi.client.Loyalty.GetCharactersCharacterIdLoyaltyPoints(
             character_id=character.eve_character.character_id,
-            token=token.valid_access_token(),
-        ).results()
-
-        return loyalty_entries
+            token=token,
+        ).result(use_etag=False)
+        return [obj.model_dump() for obj in loyalty_entries]
 
     def _update_or_create_objs(
         self, character: Character, esi_data: List[dict]
@@ -485,12 +482,18 @@ class CharacterMailManager(models.Manager):
         page = 1
         while True:
             logger.info("%s: Fetching mail headers from ESI - page %s", character, page)
-            mail_headers = esi.client.Mail.get_characters_character_id_mail(
-                character_id=character.eve_character.character_id,
-                last_mail_id=last_mail_id,
-                token=token.valid_access_token(),
-            ).results()
+            params = {
+                "character_id": character.eve_character.character_id,
+                "token": token,
+            }
+            if last_mail_id:
+                params["last_mail_id"] = last_mail_id
 
+            headers = esi.client.Mail.GetCharactersCharacterIdMail(**params).result(
+                use_etag=False
+            )
+
+            mail_headers = [obj.model_dump() for obj in headers]
             mail_headers_all += mail_headers
             if len(mail_headers) < 50 or len(mail_headers_all) >= MEMBERAUDIT_MAX_MAILS:
                 break
@@ -691,14 +694,17 @@ class CharacterMailManager(models.Manager):
         """Update or create mail body for a character from ESI."""
         try:
             mail_body = self._fetch_mail_body_from_esi(character, mail)
-        except HTTPNotFound:
-            logger.info(
-                "%s: Mail %s was deleted in game. Removing mail header.",
-                character,
-                mail,
-            )
-            mail.delete()
-            return UpdateSectionResult(is_changed=True, is_updated=True)
+        except HTTPClientError as ex:
+            if ex.status_code == HTTPStatus.NOT_FOUND:
+                logger.info(
+                    "%s: Mail %s was deleted in game. Removing mail header.",
+                    character,
+                    mail,
+                )
+                mail.delete()
+                return UpdateSectionResult(is_changed=True, is_updated=True)
+
+            raise ex
 
         is_changed = mail.body != mail_body
         if is_changed or force_update:
@@ -725,13 +731,13 @@ class CharacterMailManager(models.Manager):
             "%s: Fetching mail body from ESI for mail ID %d", character, mail.mail_id
         )
         token = character.fetch_token("esi-mail.read_mail.v1")
-        mail_body = esi.client.Mail.get_characters_character_id_mail_mail_id(
+        mail = esi.client.Mail.GetCharactersCharacterIdMailMailId(
             character_id=character.eve_character.character_id,
             mail_id=mail.mail_id,
-            token=token.valid_access_token(),
-        ).result()
+            token=token,
+        ).result(use_etag=False)
 
-        return mail_body.get("body") or ""
+        return mail.body or ""
 
 
 class CharacterMailLabelManager(models.Manager):
@@ -753,25 +759,30 @@ class CharacterMailLabelManager(models.Manager):
         )
 
     @fetch_token_for_character("esi-mail.read_mail.v1")
-    def _fetch_data_from_esi(self, character: Character, token: Token) -> dict:
+    def _fetch_data_from_esi(
+        self, character: Character, token: Token
+    ) -> Dict[int, dict]:
         from memberaudit.models import CharacterMailUnreadCount
 
         logger.info("%s: Fetching mail labels from ESI", character)
-        mail_labels_info = esi.client.Mail.get_characters_character_id_mail_labels(
+        mail_labels_info = esi.client.Mail.GetCharactersCharacterIdMailLabels(
             character_id=character.eve_character.character_id,
-            token=token.valid_access_token(),
-        ).results()
+            token=token,
+        ).result(use_etag=False)
 
-        if mail_labels_info.get("total_unread_count"):
+        if total := mail_labels_info.total_unread_count:
             CharacterMailUnreadCount.objects.update_or_create(
-                character=character,
-                defaults={"total": mail_labels_info.get("total_unread_count")},
+                character=character, defaults={"total": total}
             )
 
-        mail_labels = mail_labels_info.get("labels")
-        if not mail_labels:
+        if not mail_labels_info.labels:
             return {}
-        return {obj["label_id"]: obj for obj in mail_labels if "label_id" in obj}
+
+        return {
+            obj.label_id: obj.model_dump()
+            for obj in mail_labels_info.labels
+            if obj.label_id
+        }
 
     @transaction.atomic()
     def _update_or_create_objs(self, character: Character, mail_labels_list: dict):
